Extract code_for helper into CgminerManager::ErrorCode module
`AdminLogging`'s class-level header documents it as "pure helpers for
the admin-surface plumbing: session-id hashing and structured-log-entry
construction." `code_for` was neither — it's a generic
`CgminerApiClient`-error → symbol classifier with no admin-surface
coupling, and `FleetQueryResult` / `FleetWriteResult` reached back into
`AdminLogging` purely to compute a value those Data classes own.

Move it to a new `CgminerManager::ErrorCode` module exposed as
`.classify(error)`. Renamed along the way: `def for(error)` would have
parse-errored as a Ruby reserved-word landmine, and `.for(error)` reads
ambiguously next to `CgminerApiClient::ApiError#code` ("construct an
ErrorCode value?"). `.classify(error)` reads exactly like what it does.

Internal refactor — no behavior change, no log-shape change. Two
in-repo callsites updated; spec coverage relocated to a dedicated
`error_code_spec.rb`. Bumps to v1.6.2.
