+++ lib/cgminer_manager.rb
@@ -2,6 +2,7 @@
 
 require_relative 'cgminer_manager/version'
 require_relative 'cgminer_manager/errors'
+require_relative 'cgminer_manager/error_code'
 require_relative 'cgminer_manager/config'
 require_relative 'cgminer_manager/logger'
 require_relative 'cgminer_manager/monitor_client'
