+++ lib/cgminer_manager/fleet_write_result.rb
@@ -21,7 +21,7 @@ module CgminerManager
     # cgminer_monitor's docs/log_schema.md `code` row.
     def failed_codes_count_map
       entries.select(&:failed?).each_with_object(Hash.new(0)) do |entry, counts|
-        counts[CgminerManager::AdminLogging.code_for(entry.error)] += 1
+        counts[CgminerManager::ErrorCode.classify(entry.error)] += 1
       end
     end
   end
